refactor(it): use PackageMarker instead of referencing all classes
