+++ web/src/main/java/io/fabric8/launcher/web/PackageMarker.java
@@ -1,0 +1,7 @@
+package io.fabric8.launcher.web;
+
+/**
+ * This interface helps reference this package (ex: for CDI).
+ */
+public interface PackageMarker {
+}
