+++ web/src/test/java/io/fabric8/launcher/web/endpoints/websocket/MissionControlStatusEndpointIT.java
@@ -12,7 +12,7 @@ import javax.ws.rs.core.UriBuilder;
 
 import io.fabric8.launcher.core.api.events.StatusEventType;
 import io.fabric8.launcher.core.api.events.StatusMessageEvent;
-import io.fabric8.launcher.web.endpoints.Deployments;
+import io.fabric8.launcher.web.Deployments;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
